refact :: [#108] model 한 파일에 넣기 및 위치 조정 그룹화
